Use delegated volume mount to increase performance on macOS.
Newer versions of Docker for Mac support caching options for macOS
volume mounts. Using the `delegated` mount option allows the container
and host to temporarily be out of sync, making the container's view
authoritative, which can improve performance because writes within the
container do not need to propagate to the host before they return.

Empirically, this slightly speed up the Webpack build from ~40s to ~35s
on my MacBook Pro.

https://docs.docker.com/compose/compose-file/#caching-options-for-volume-mounts-docker-for-mac